+++ src/sensors/HMC5883L.h
@@ -1,6 +1,7 @@
 #ifndef APOLLON_FC_SRC_SENSORS_HMC5883L_H
 #define APOLLON_FC_SRC_SENSORS_HMC5883L_H
 
+#include <Arduino.h>
 #include "../sensor.h"
 
 /* HMC5883L Sensor
@@ -46,14 +47,14 @@ typedef enum : byte { // 3-bit gain value
   HMC5883L_GAIN_7 = 0x7,
 } hmc5883l_gain_t;
 
-typedef enum : byte { // 2-bits
+typedef enum : byte { // 2-bit oversampling setting
   HMC5883L_OSS_0 = 0x0,
   HMC5883L_OSS_1 = 0x1,
   HMC5883L_OSS_2 = 0x2,
   HMC5883L_OSS_3 = 0x3,
 } hmc5883l_oss_t;
 
-typedef enum : byte { // 3-bits
+typedef enum : byte { // 3-bit output rate setting (continuous modde only)
   HMC5883L_OUTRATE_0 = 0x0,
   HMC5883L_OUTRATE_1 = 0x1,
   HMC5883L_OUTRATE_2 = 0x2,
@@ -64,7 +65,7 @@ typedef enum : byte { // 3-bits
   HMC5883L_OUTRATE_7 = 0x7,
 } hmc5883l_outrate_t;
 
-typedef enum : byte { // 2-bits, excluding 0b11
+typedef enum : byte { // 2-bits, excluding 0b11, bias setting (see datasheet)
   HMC5883L_BIAS_0 = 0x0,
   HMC5883L_BIAS_1 = 0x1,
   HMC5883L_BIAS_2 = 0x2,
@@ -72,26 +73,9 @@ typedef enum : byte { // 2-bits, excluding 0b11
 
 
 class Sensor_HMC5883L : I2C_Sensor {
-private:
-  byte lock; // Status bit 1, write locks output registers
-  byte ready; // Status bit 0, high when measurements are ready
-  hmc5883l_gain_t gain; // Higher gain # decreases actual gain and increases signal range
-  const hmc5883l_mode_t mode;
-  static constexpr int16_t gainLSB_Gauss[] = {1370, 1090, 820, 660, 440, 390, 330, 230};
 public:
   int16_t x = 0, y = 0, z = 0; // magnetometer values in microGauss
 
-private:
-  /* Read status register for lock and ready bits */
-  void readStatus(const bool closeConnection);
-
-  /* Sets mode register, which begins measurement
-   * For continuous-mode operation, this only needs to be called once
-   * For single-measurement operation, this needs to be called after data is read
-   */
-  void startMeasurement(const bool closeConnection);
-
-public:
   Sensor_HMC5883L(
     const hmc5883l_mode_t _mode = HMC5883L_MODE_SINGLE,
     const hmc5883l_gain_t _gain = HMC5883L_GAIN_1,
@@ -103,75 +87,22 @@ public:
   void setGain(const hmc5883l_gain_t g, const bool closeConnection);
 
   bool update();
+
+private:
+  byte lock; // Status bit 1, write locks output registers
+  byte ready; // Status bit 0, high when measurements are ready
+  hmc5883l_gain_t gain; // Higher gain # decreases actual gain and increases signal range
+  const hmc5883l_mode_t mode;
+  static constexpr int16_t gainLSB_Gauss[] = {1370, 1090, 820, 660, 440, 390, 330, 230};
+
+  /* Read status register for lock and ready bits */
+  void readStatus(const bool closeConnection);
+
+  /* Sets mode register, which begins measurement
+   * For continuous-mode operation, this only needs to be called once
+   * For single-measurement operation, this needs to be called after data is read
+   */
+  void startMeasurement(const bool closeConnection);
 };
-constexpr int16_t Sensor_HMC5883L::gainLSB_Gauss[]; // Fixed undefined reference error
-
-
-
-/* Implementation */
-
-Sensor_HMC5883L::Sensor_HMC5883L(
-    const hmc5883l_mode_t _mode = HMC5883L_MODE_SINGLE,
-    const hmc5883l_gain_t _gain = HMC5883L_GAIN_1,
-    const hmc5883l_oss_t oss = HMC5883L_OSS_0,
-    const hmc5883l_outrate_t outRate = HMC5883L_OUTRATE_4,
-    const hmc5883l_bias_t bias = HMC5883L_BIAS_0
-) : I2C_Sensor(HMC5883L_ADDRESS), mode(_mode) {
-  byte id[3];
-
-  // Check ID = H43
-  selectRegister(HMC5883L_REGISTER_IDA, connection.keepAlive);
-  readIntoArray(id, 3, connection.close);
-  if (id[0] != 0x48 || id[1] != 0x34 || id[2] != 0x33)
-    ERROR(3); // I2C configuration error
-
-  // Configure device
-  setGain(_gain, connection.keepAlive);
-  write8(HMC5883L_REGISTER_CFGA, (oss << 5) | (outRate << 3) | bias, connection.keepAlive);
-  startMeasurement(connection.close);
-}
-
-void Sensor_HMC5883L::readStatus(const bool closeConnection) {
-  byte status;
-  selectRegister(HMC5883L_REGISTER_STATUS, connection.keepAlive);
-  read8(status, closeConnection);
-  lock = status & 0x02;
-  ready = status & 0x01;
-}
-
-void Sensor_HMC5883L::startMeasurement(const bool closeConnection) {
-  write8(HMC5883L_REGISTER_MODE, (HMC5883L_HIGH_SPEED << 7) | mode, closeConnection);
-}
-
-void Sensor_HMC5883L::setGain(const hmc5883l_gain_t g, const bool closeConnection) {
-  gain = g;
-  write8(HMC5883L_REGISTER_CFGB, gain << 5, closeConnection);
-}
-
-bool Sensor_HMC5883L::update() {
-  int16_t data[3];
-
-  readStatus(connection.close); // Close connection in case not ready
-  if (ready) {
-    // Read magnetometer data
-    selectRegister(HMC5883L_REGISTER_X_MSB, connection.keepAlive);
-    readIntoArray(data, 3, connection.keepAlive);
-    startMeasurement(connection.close); // Start next measurement cycle
-
-    // If any value is -4096, it has overflowed
-    if (data[0] == -4096 || data[1] == -4096 || data[2] == -4096) {
-      Serial.println("WARNING HMC5883L: integer overflow detected (magnetic field outside sensor range). Try a higher gain value if this problem persists.");
-      return false;
-    }
-
-    // Convert from count (LSB) to microGauss measurements
-    x = ( (int32_t) data[0] * 1000) / gainLSB_Gauss[gain];
-    z = ( (int32_t) data[1] * 1000) / gainLSB_Gauss[gain];
-    y = ( (int32_t) data[2] * 1000) / gainLSB_Gauss[gain];
-
-    return true;
-  }
-  return false;
-}
 
 #endif
